refactor: weather, forecast viewmodel, usecase, repository 하나로 합치기
