--- WeatherForecast/WeatherForecast/Domain/Interfaces/Repositories/ForecastRepositoryInterface.swift
@@ -1,12 +1,0 @@
-//
-//  ForecastRepositoryInterface.swift
-//  WeatherForecast
-//
-//  Created by devxsby on 2023/03/30.
-//
-
-import Foundation
-
-protocol ForecastRepositoryInterface {
-    func fetchForecast(lat: String, lon: String, completion: @escaping (ForecastEntity) -> Void)
-}
