+++ WeatherForecast/WeatherForecast/Domain/Interfaces/Repositories/WeatherForecastRepositoryInterface.swift
@@ -1,0 +1,18 @@
+//
+//  WeatherForecastRepositoryInterface.swift
+//  WeatherForecast
+//
+//  Created by devxsby on 2023/04/03.
+//
+
+import Foundation
+
+protocol WeatherForecastRepositoryInterface: WeatherRepositoryInterface, ForecastRepositoryInterface { }
+
+protocol WeatherRepositoryInterface {
+    func fetchWeather(lat: String, lon: String, completion: @escaping(WeatherEntitiy) -> Void)
+}
+
+protocol ForecastRepositoryInterface {
+    func fetchForecast(lat: String, lon: String, completion: @escaping (ForecastEntity) -> Void)
+}
