--- WeatherForecast/WeatherForecast/Domain/Interfaces/Repositories/WeatherRepositoryInterface.swift
@@ -1,12 +1,0 @@
-//
-//  WeatherRepositoryInterface.swift
-//  WeatherForecast
-//
-//  Created by devxsby on 2023/03/30.
-//
-
-import Foundation
-
-protocol WeatherRepositoryInterface {
-    func fetchWeather(lat: String, lon: String, completion: @escaping(WeatherEntitiy) -> Void)
-}
